mandatory done && functions work

+++ libft/ft_nbrlen.c
@@ -6,20 +6,20 @@
 /*   By: edpaulin <edpaulin@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/20 14:22:07 by edpaulin          #+#    #+#             */
-/*   Updated: 2021/08/20 17:47:16 by edpaulin         ###   ########.fr       */
+/*   Updated: 2021/08/21 12:37:54 by edpaulin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/ft_printf.h"
 
-int	ft_nbrlen(ssize_t	n, int	base)
+int	ft_nbrlen(size_t	n, size_t	base)
 {
 	int	len;
 
 	len = 0;
 	while (n)
 	{
-		n /= (ssize_t)base;
+		n /= base;
 		++len;
 	}
 	return (len);
