+++ srcs/ft_type_x_lower.c
@@ -6,7 +6,7 @@
 /*   By: edpaulin <edpaulin@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/20 17:35:26 by edpaulin          #+#    #+#             */
-/*   Updated: 2021/08/20 17:54:33 by edpaulin         ###   ########.fr       */
+/*   Updated: 2021/08/21 13:22:06 by edpaulin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 int	ft_type_x_lower(va_list	ap)
 {
-	size_t	nbr;
-	int		len;
-	char	*hex_nbr;
+	unsigned int	nbr;
+	int				len;
+	char			*hex_nbr;
 
-	nbr = va_arg(ap, size_t);
-	len = 0;	
+	nbr = va_arg(ap, unsigned int);
+	len = 0;
 	hex_nbr = ft_convert_base(nbr, LOWER_HEX);
 	if (hex_nbr)
 	{
